requirements.txt
flask==3.0.3
openai==1.35.13
httpx==0.27.0
PyMuPDF==1.23.26
pillow==10.4.0
python-dotenv==1.0.1